refactor: cleanup nav component event listeners
