+++ src/components/index.js
@@ -2,3 +2,4 @@ export * from "./header";
 export * from "./nav";
 export * from "./footer";
 export * from "./submit-button";
+export * from "./post-card";
