+++ src/pages/(app)/landing/page.js
@@ -1,7 +1,16 @@
 import { landingTemplate } from "./page.template";
+import Layout from "../layout";
 
 export default function LandingPage(container) {
   if (!container) return null;
 
-  container.innerHTML = landingTemplate();
+  const { html, cleanUp } = Layout(container, {
+    children: landingTemplate,
+  });
+
+  container.innerHTML = html;
+
+  return () => {
+    if (cleanUp) cleanUp();
+  };
 }
