+++ src/pages/(app)/layout.js
@@ -1,7 +1,25 @@
-import { Header, Nav, Footer } from "../../components";
+import env from "@lib/contexts/env";
+import { Header, Nav, Footer } from "@components";
 
-export default function Layout(children) {
-  return `
+export default function Layout(container, { children }) {
+  if (!container) return null;
+
+  const handleNavigation = (e) => {
+    if (e.target && e.target.matches('nav a:not([id="logout"])')) {
+      e.preventDefault();
+      const href = e.target.getAttribute("href");
+      env.router.navigate(href);
+    }
+  };
+
+  const handleLogout = (e) => {
+    if (e.target && e.target.id === "logout") {
+      e.preventDefault();
+      env.authService.logout();
+    }
+  };
+
+  const html = `
     <div class="bg-gray-100 min-h-screen flex justify-center">
       <div class="max-w-md w-full">
         ${Header({ title: "항해플러스" })}
@@ -13,4 +31,17 @@ export default function Layout(children) {
       </div>
     </div>
   `;
+
+  const cleanUp = () => {
+    document.removeEventListener("click", handleNavigation);
+    document.removeEventListener("click", handleLogout);
+  };
+
+  document.addEventListener("click", handleNavigation);
+  document.addEventListener("click", handleLogout);
+
+  return {
+    html,
+    cleanUp,
+  };
 }
